Cleaned up design
Separate storyboard for data history was removed, and simply added as a view controller to the dashboard storyboard.
Segues are now coming from the DashboardVC to the DataHistoryVC and to the TOot...MasterVC. The correct segue is performed inside tableView(didSelectRowAt:), and prep is done inside prepare(forSegue:)
This allows a cleaner design, as well as the ability to embed the SAP generated IOT VC's into the Navigation View Controller, and a 'Back' button is now available.