Connected our data to the chart
Changed data back to array of array of Double, and during the data fetch add the data into this array, and store it inside the cardChart variable as a chart data structure.,
In order for the tableview to correctly display the data, a switch statement is added into the    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell which houses the logic of the conversion to a displayable format.

Since data is not ready initially, a reload of the chart card (collection view) is added to the main operation queue if it finds the data is empty.